Minor changes for primitive related tests.
Instead of using equals method, the test will compare each field manually. This way, if a problem occurs with saving and loading a field, we will be able to spot the exact field where the problem occurred.

Additionally, after the functionality change in commit fe5148f8cd67ba23b5d79c221cfa3bbf63ac8155, it seems like the type of the field is always a wrapper class. This is likely due to getting the type with value using getClass.